Tweak local build scripts

+++ _build/copy_artifacts.ps1
@@ -6,5 +6,7 @@ Import-Csv .\_build\files.csv | ForEach-Object {
     $s = "{0}{1}" -f $_.sourcePath,$buildConfiguration
     $p = Join-Path -Path $s -ChildPath $_.fileName 
     $d = "{0}\{1}" -f $addinFolder,$_.destination
-    Copy-Item $p -Destination $d
+    If ((Test-Path $p)) {
+        Copy-Item $p -Destination $d
+    }
 }
